--- a/packages/env/lib/wordpress.js
+++ b/packages/env/lib/wordpress.js
@@ -94,10 +94,6 @@ async function configureWordPress( environment, config, spinner ) {
 		setupCommands.push( `wp plugin activate ${ pluginSource.basename }` );
 	}
 
-	setupCommands.push(
-		'sed "/^require.*wp-settings.php/d" /var/www/html/wp-config.php > /var/www/html/phpunit-wp-config.php'
-	);
-
 	if ( config.debug ) {
 		spinner.info(
 			`Running the following setup commands on the ${ environment } instance:\n - ${ setupCommands.join(
@@ -116,23 +112,31 @@ async function configureWordPress( environment, config, spinner ) {
 		}
 	);
 
-	// // Since wp-phpunit loads wp-settings.php at the end of its wp-config.php
-	// // file, we need to avoid loading it too early in our own wp-config.php. If
-	// // we load it too early, then some things (like MULTISITE) will be defined
-	// // before wp-phpunit has a chance to configure them. To avoid this, create a
-	// // copy of wp-config.php for phpunit which doesn't require wp-settings.php.
-	// await dockerCompose.exec(
-	// 	environment === 'development' ? 'wordpress' : 'tests-wordpress',
-	// 	[
-	// 		'sh',
-	// 		'-c',
-	// 		'sed "/^require.*wp-settings.php/d" /var/www/html/wp-config.php > /var/www/html/phpunit-wp-config.php',
-	// 	],
-	// 	{
-	// 		config: config.dockerComposeConfigPath,
-	// 		log: config.debug,
-	// 	}
-	// );
+	/**
+	 * Since wp-phpunit loads wp-settings.php at the end of its wp-config.php
+	 * file, we need to avoid loading it too early in our own wp-config.php. If
+	 * we load it too early, then some things (like MULTISITE) will be defined
+	 * before wp-phpunit has a chance to configure them. To avoid this, create a
+	 * copy of wp-config.php for phpunit which doesn't require wp-settings.php.
+	 *
+	 * Note that This needs to be executed using `exec` on the wordpress service
+	 * so that file permissions work properly.
+	 *
+	 * This will be removed in the future. @see https://github.com/WordPress/gutenberg/issues/23171
+	 *
+	 */
+	await dockerCompose.exec(
+		environment === 'development' ? 'wordpress' : 'tests-wordpress',
+		[
+			'sh',
+			'-c',
+			'sed "/^require.*wp-settings.php/d" /var/www/html/wp-config.php > /var/www/html/phpunit-wp-config.php',
+		],
+		{
+			config: config.dockerComposeConfigPath,
+			log: config.debug,
+		}
+	);
 }
 
 /**