<!-- Learn the overall process and best practices for pull requests at https://github.com/WordPress/gutenberg/blob/master/docs/contributors/repository-management.md#pull-requests. -->

## Description
Resolves #23175. Improves startup performance by not re-launching the docker service over and over again. Rather, we run all the commands in bash with a single service execution.

Additionally, I removed the behavior which activates the first theme in the list. I personally think this is ok to do since it's really annoying to have your chosen theme be overwritten every time you start wp-env.

## How has this been tested?
Locally in wp-env; CI should pass

## Performance improvement

Each value below is the fastest run time out of 5 runs.

|            | Fresh Install* | Fresh Start | Restart (already running) |
| ---------- | ------------- | ----------- | ------------------------- |
| Previously | 63 seconds    | 24 seconds  | 29 seconds                |
| Now 🔥     | 49 seconds    | 10 seconds  | 13 seconds                |

In general, we see that wp-env runs around 14 seconds faster, which about twice as fast in normal use cases. We can further improve this by avoiding WordPress configuration when nothing has changed in the config files.

* The performance of a fresh install is mostly affected by your internet connection for downloading WordPress.

## Types of changes

## Checklist:
- [ ] My code is tested.
- [x] My code follows the WordPress code style. <!-- Check code: `npm run lint`, Guidelines: https://make.wordpress.org/core/handbook/best-practices/coding-standards/javascript/ -->
- [x] My code follows the accessibility standards. <!-- Guidelines: https://make.wordpress.org/core/handbook/best-practices/coding-standards/accessibility-coding-standards/ -->
- [ ] My code has proper inline documentation. <!-- Guidelines: https://make.wordpress.org/core/handbook/best-practices/inline-documentation-standards/javascript/ -->
- [x] I've included developer documentation if appropriate. <!-- Handbook: https://developer.wordpress.org/block-editor/ -->
- [x] I've updated all React Native files affected by any refactorings/renamings in this PR. <!-- React Native mobile Gutenberg guidelines: https://github.com/WordPress/gutenberg/blob/master/docs/contributors/native-mobile.md -->