@epiqueras 2020-07-08
There is a permissions issue with `phpunit-wp-config`.