@noahtallen 2020-07-08
Currently failing in CI with: `bash: /var/www/html/phpunit-wp-config.php: Permission denied`

which is interesting because it worked before 🤔 possible changes affecting it:
1. run with bash not sh
2. run with ci container, not wordpress
3. run with docker run, not docker exec