@noahtallen 2020-07-09
Everything should be fixed now. We'll just leave the command for phpunit config the same since it might be removed with #23171